DiscordOnlineTracker/requirements.txt
discord.py>=2.6.4
flask>=3.1.2
requests>=2.32.5
